+++ src/coal_plant.cpp
@@ -11,7 +11,7 @@ public:
 
     int get_emissionsOutput(Emissions& emissions) override;
 
-    int get_productionCost(float &cost, float powerRequest) override;
+    int get_productionCost(double &cost, double powerRequest) override;
 
 private:
     CoalPlant(const CoalPlant&) = delete;
@@ -40,7 +40,7 @@ CoalPlant::get_emissionsOutput(Emissions& emissions)
 
 
 int
-CoalPlant::get_productionCost(float &cost, float powerRequest)
+CoalPlant::get_productionCost(double &cost, double powerRequest)
 {
     cost = powerRequest * _runCost / 3600;
 }
